rename Modal --> Popup
also adds a bunch of fixes
